Fix #4531 gap 2: the local gate runs the reachability walk
Done: harness_report and its helpers extracted to
scripts/ci/harness_reachability.py; imported by both the test module and
validate_agent_setup.py, which now reports orphans, broken links, stale named
paths and invalid exemptions.

Mutation-proven, not asserted. An unreferenced file under
.agents/skills/act-as-mohab/references/ makes the command exit 1 and name the
path; removing it returns exit 0. Before this, the command AGENTS.md tells
agents to run exited 0 on a fully orphaned harness, so only a GitHub Actions
job ever reported orphans -- not a guarantee for a machine that never opens a
pull request, which is the requirement.

Two things the unit tests could not have caught, both found by running the
command the way an agent runs it:

- The import was placed above the sys.path insert. All 51 tests passed,
  because unittest runs from the repository root where `scripts` is already
  importable, while the command died with ModuleNotFoundError from any other
  cwd -- the one invocation this function exists to serve. Moved below, with
  the sibling imports.
- The first mutation proved nothing: severing the roles.md link left exit 0,
  correctly, because reachability also flows through the skills map and
  parallel entrypoint links are redundant by design. Reading that as a defect
  would have chased a phantom; reading the passing tests as proof would have
  shipped a broken command.

Duplicating the walk was rejected on this repository's own evidence: #4481
recorded four of five copies of a references glob failing open after the path
moved. Every assertion and mutation fixture stays in the test module.

Checks: harness_reachability + validate_agent_setup + harness_portability +
router_contract + validate_agent_guidance 231 OK; the command run as a script
from an unrelated cwd exits 0 clean and 1 on a seeded orphan.

Remaining: #4531 gap 3; #4541 registry, now targeting a mechanism for all 12
required-action rules rather than a prose-only category; retrieval skill;
#4529, #4532, #4534, #4535.

Blocking: none.
